refactor: chat - css 정리
