+++ src/routes/chat/Chat.css
@@ -1,76 +1,49 @@
-img{ max-width:100%;}
-
 .chat_container {
   background-color: white;
   /* clear: both; */
   overflow: hidden;
   padding: 12px;
 }
+
+/* 왼쪽 섹션 시작 */
 .meetups_box {
-  background: #f8f8f8 none repeat scroll 0 0;
+  /* background: #f8f8f8 none repeat scroll 0 0; */
   float: left;
   overflow: hidden;
   width: 40%; 
-  border-right:1px solid #c4c4c4;
 }
 .meetups_box_header { 
   padding:10px 0 10px 0; 
   overflow:hidden; 
   border-bottom:1px solid #c4c4c4;
-}
-
-.top_spac { 
-  margin: 20px 0 0;
-}
-/* .recent_heading {float: left; width:40%;} */
-/* .srch_bar {
-  display: inline-block;
-  text-align: right;
-  width: 60%;
-} */
-.recent_heading h4 {
   color: #0584BB;
   margin: auto;
 }
-.srch_bar input{ 
-  border:1px solid #cdcdcd; 
-  border-width:0 0 1px 0; 
-  width:80%; 
-  padding:2px 0 4px 6px; 
-  background:none;
-}
-.srch_bar .input-group-addon button {
-  background: rgba(0, 0, 0, 0) none repeat scroll 0 0;
-  border: medium none;
-  padding: 0;
-  color: #707070;
-  font-size: 18px;
-}
-.srch_bar .input-group-addon { 
-  margin: 0 0 0 -27px;
+.meetups_box_body { 
+  height: 530px; 
+  overflow-y: scroll;
 }
 
-
-.chat_item {
+.meetups_box_body_item {
   border-bottom: 1px solid #c4c4c4;
   margin: 0;
   padding: 10px;
   overflow:hidden; 
   clear:both;
   width: 100%; 
-  height: 100px;
+  height: 97px;
 }
 
-.chat_item h5 { 
+.meetups_box_body_item h5 { 
   font-size:15px; 
   color:#464646; 
   margin:0 0 8px 0;
 }
-.chat_item h5 span { 
+.meetups_box_body_item h5 span { 
   font-size:13px; 
   float:right;
 }
-.chat_item p { 
+.meetups_box_body_item p { 
   width: 80%;
   float: left;
   font-size:14px; 
@@ -83,50 +56,51 @@ img{ max-width:100%;}
   -webkit-line-clamp: 2;
   -webkit-box-orient: vertical;
 }
+.active { 
+  background:#ebebeb; 
+}
+/* 왼쪽 섹션 끝 */
 
-.meetups_box_body { 
-  height: 530px; 
-  overflow-y: scroll;
-}
-
-.active_chat { background:#ebebeb; }
-/* 
-.incoming_msg_img {
-  display: inline-block;
-  width: 15%;
-} */
-.received_msg {
-  display: inline-block;
-  /* padding: 0 0 0 10px; */
-  vertical-align: top;
-  width: 100%;
-}
-.received_withd_msg { 
-  width: 57%;
-}
-.received_withd_msg p {
-  background: #ebebeb none repeat scroll 0 0;
-  border-radius: 3px;
-  color: #646464;
-  font-size: 14px;
-  margin: 0;
-  padding: 5px 10px 5px 12px;
-  width: 100%;
-}
-.msg_date {
-  color: #747474;
-  display: block;
-  font-size: 12px;
-  margin: 8px 0 0;
-}
-
+/* 오른쪽 섹션 시작 */
 .chat_box {
   float: left;
   padding: 30px 15px 0 25px;
   width: 60%;
 }
+.chat_history {
+  height: 500px;
+  overflow-y: auto;
+}
 
-.sent_msg p {
+/* 상대 메시지 */
+.incoming_msg_writer {
+  font-size: 15px;
+  color: #464646;
+}
+.incoming_msg_content {
+  display: inline-block;
+  vertical-align: top;
+  width: 55%;
+}
+.incoming_msg_content p {
+  background: #ebebeb none repeat scroll 0 0;
+  border-radius: 3px;
+  font-size: 14px;
+  margin: 0;
+  padding: 5px 10px 5px 12px;
+  width: 100%;
+}
+
+/* 내 메시지 */
+.outgoing_msg { 
+  overflow:hidden; 
+  margin:26px 0 26px; 
+}
+.outgoing_msg_content {
+  float: right;
+  width: 55%;
+}
+.outgoing_msg_content p {
   background: #0584BB none repeat scroll 0 0;
   border-radius: 3px;
   font-size: 14px;
@@ -134,31 +108,29 @@ img{ max-width:100%;}
   padding: 5px 10px 5px 12px;
   width:100%;
 }
-.outgoing_msg { 
-  overflow:hidden; 
-  margin:26px 0 26px; 
+
+.msg_date {
+  color: #747474;
+  display: block;
+  font-size: 12px;
+  margin: 8px 0 0;
 }
-.sent_msg {
-  float: right;
-  width: 46%;
+
+/* 채팅 입력 */
+.write_msg { 
+  border-top: 1px solid #c4c4c4;
+  position: relative; 
 }
-.input_msg_write input {
-  background: rgba(0, 0, 0, 0) none repeat scroll 0 0;
+.write_msg input {
+  /* background: rgba(0, 0, 0, 0) none repeat scroll 0 0; */
   border: medium none;
   color: #4c4c4c;
   font-size: 15px;
-  min-height: 48px;
+  min-height: 43px;
   width: 100%;
 }
-
-.type_msg { border-top: 1px solid #c4c4c4;position: relative; }
-
 .chatSendBtn {
   position: absolute;
-  right: 0;
-  top: 11px;
-}
-.msg_history {
-  height: 500px;
-  overflow-y: auto;
+  right: 4px;
+  top: 6px;
 }
